+++ data/src/main/java/vzijden/workout/data/model/PlannedExercisePojo.kt
@@ -3,18 +3,19 @@ package vzijden.workout.data.model
 import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import androidx.room.Relation
 
 @Entity
 class PlannedExercisePojo(
-    var workoutId: Long,
+    var workoutId: Int,
     @Embedded(prefix = "exercise_")
     var exercisePojo: ExercisePojo,
     var index: Int
 ) {
-  constructor(workoutId: Long, exercisePojo: ExercisePojo, index: Int, id: Long) : this(workoutId, exercisePojo, index) {
+  constructor(workoutId: Int, exercisePojo: ExercisePojo, index: Int, id: Int) : this(workoutId, exercisePojo, index) {
     this.id = id
   }
 
   @PrimaryKey(autoGenerate = true)
-  var id: Long = 0
+  var id: Int = 0
 }
