+++ domain/src/main/java/vzijden/workout/domain/usecase/GetPlannedWorkouts.kt
@@ -9,10 +9,10 @@ import vzijden.workout.domain.repository.WorkoutRepository
 
 class GetPlannedWorkouts(private val workoutRepository: WorkoutRepository, subscribeScheduler: Scheduler,
                          postExecutionScheduler: Scheduler) :
-    ObservableUseCase<List<PlannedWorkout>, Long>(subscribeScheduler, postExecutionScheduler) {
+    ObservableUseCase<List<PlannedWorkout>, Int>(subscribeScheduler, postExecutionScheduler) {
 
 
-  override fun build(scheduleId: Long?): Observable<List<PlannedWorkout>> {
+  override fun build(scheduleId: Int?): Observable<List<PlannedWorkout>> {
     return workoutRepository.getPlannedWorkouts(scheduleId!!)
   }
 }
