+++ domain/src/main/java/vzijden/workout/domain/usecase/StartWorkout.kt
@@ -8,9 +8,9 @@ import vzijden.workout.domain.repository.WorkoutRepository
 import java.util.*
 
 class StartWorkout(private val workoutRepository: WorkoutRepository, subscribeScheduler: Scheduler, postExecutionScheduler: Scheduler) :
-    SingleUseCase<Long, Int>(subscribeScheduler, postExecutionScheduler) {
+    SingleUseCase<Int, Int>(subscribeScheduler, postExecutionScheduler) {
 
-  override fun build(workoutId: Int): Single<Long> {
+  override fun build(workoutId: Int): Single<Int> {
     val loggedWorkout = LoggedWorkout(Date(), workoutId)
     return workoutRepository.saveLoggedWorkout(loggedWorkout)
   }
