+++ domain/src/main/java/vzijden/workout/domain/usecase/UpdateWorkout.kt
@@ -7,9 +7,9 @@ import vzijden.workout.domain.model.PlannedWorkout
 import vzijden.workout.domain.repository.WorkoutRepository
 
 class UpdateWorkout(private var workoutRepository: WorkoutRepository, subscribeScheduler: Scheduler, postExecutionScheduler: Scheduler) :
-    SingleUseCase<Long, PlannedWorkout>(subscribeScheduler, postExecutionScheduler) {
+    SingleUseCase<Int, PlannedWorkout>(subscribeScheduler, postExecutionScheduler) {
 
-  override fun build(plannedWorkout: PlannedWorkout): Single<Long> {
+  override fun build(plannedWorkout: PlannedWorkout): Single<Int> {
     return workoutRepository.savePlannedWorkout(plannedWorkout)
   }
 }
