+++ mobile/src/main/java/vzijden/workout/view/home/MainActivity.kt
@@ -38,7 +38,7 @@ class MainActivity : DaggerAppCompatActivity() {
         Log.d(App.TAG, "initializing database")
         initialize(workoutDatabase, this@MainActivity.resources.openRawResource(R.raw.exercises))
         val scheduleId = workoutDatabase.scheduleDao().insert(schedule)
-        val workoutId = workoutDatabase.workoutDao().insert(PlannedWorkoutPojo(scheduleId.toInt(), "Chest", 1)).blockingGet()
+        val workoutId = workoutDatabase.workoutDao().insert(PlannedWorkoutPojo(scheduleId, "Chest", 1)).blockingGet()
         val exercises = workoutDatabase.exerciseDao().getAll().blockingFirst()
         val registrationId = workoutDatabase.registrationDao().insert(PlannedExercisePojo(workoutId, exercises[0], 0)).blockingGet()
         workoutDatabase.setsDao().insert(PlannedSetPojo(8, 0, registrationId)).blockingGet()
